Remove dependencies from repository
Remove the dependencies from the repository, and instead provide
`cpanfile` and `cpanfile.snapshot` along with build instructions in the
readme file; this allows us to download the latest versions of
dependencies, and avoids unneeded files in the repository.

Algorithm::AM in particular needed to be updated, since a Windows crash
was fixed there recently.